pack: move to repository package
