+++ internal/repository/packer_manager.go
@@ -16,7 +16,7 @@ import (
 	"github.com/restic/restic/internal/crypto"
 	"github.com/restic/restic/internal/debug"
 	"github.com/restic/restic/internal/fs"
-	"github.com/restic/restic/internal/pack"
+	"github.com/restic/restic/internal/repository/pack"
 
 	"github.com/minio/sha256-simd"
 )
